Address PR review comments (ewoutkramer + Copilot)
Review feedback from ewoutkramer:
- Rename IrCall to IrInvoke: it invokes pre-existing .NET methods, as
  opposed to IrDefinitionCall which invokes the C# methods this compiler
  generates from CQL definitions; the names were too alike. Both classes
  and IrThrow now live in their own files.
- IrInvoke supports null-conditional invocation (x?.Method(...)), value
  return types lifted to Nullable, mirroring IrProperty - so the IR does
  not need extending when the builder ports constructs that guard method
  calls, not just member accesses.
- The emitter Scope class moved to CSharpIrEmitter.Scope.cs.

Copilot findings, all confirmed:
- Parameter name hints are now validated: hints that are C# keywords or
  collide with a name already used in the emission fall back to generated
  names (previously produced CS0100/CS0136 in generated code).
- Naming state is reset per EmitBodyBlock call instead of accumulating
  across emissions.
- Boxed default struct values (e.g. default(DateTime)) print as
  "default" instead of a non-compiling ToString rendering, restoring the
  previous pipeline behavior.
- Removed the dead NullableTestCondition helper.
- IrDefault doc now matches the actual printing contract.

Writing the null-conditional test exposed one more latent printer gap of
the same family: char constants printed bare (x) instead of as literals
(single-quoted x); now formatted via SymbolDisplay.
