+++ Cql/Cql.Compiler/Ir/IrThrow.cs
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2026, Firely, NCQA and contributors
+ * See the file CONTRIBUTORS for details.
+ *
+ * This file is licensed under the BSD 3-Clause license
+ * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
+ */
+
+namespace Hl7.Cql.Compiler.Ir;
+
+/// <summary>
+/// Throws an exception. Used only for the generator's error-recovery stubs; carries a result
+/// type so it can stand in for an expression of any type.
+/// </summary>
+internal sealed class IrThrow : IrExpression
+{
+    public IrThrow(IrExpression exception, Type resultType)
+    {
+        if (!typeof(Exception).IsAssignableFrom(exception.Type))
+            throw new ArgumentException($"Throw operand must be an Exception, not {exception.Type}.");
+        Exception = exception;
+        Type = resultType;
+    }
+
+    public IrExpression Exception { get; }
+
+    public override Type Type { get; }
+}
